requirements.txt
flask>=3.1.3
webauthn>=2.7.1,<3.0
psutil>=7.2.2
python-dotenv>=1.2.2
pyyaml>=6.0.3
